+++ Classes/AgentManager/RelativityAgentCreateResponse.ps1
@@ -1,0 +1,9 @@
+class RelativityAgentCreateResponse {
+    [ValidateNotNull()]
+    [Int32] $ArtifactID
+
+    RelativityAgentCreateResponse([Int32] $artifactID)
+    {
+        $this.ArtifactID = $artifactID
+    }
+}