+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [0.0.1] - 2023-10-31
 
 ### Added
+- AgentManager API : New-RelativityAgent
 - AgentManager API : Get-RelativityAgentType
 - ARM API : New-RelativityArmArchiveJob
 - ARM API : Get-RelativityArmArchiveJob